Options definitions inside Options class
