+++ src/Runners/PHPUnit/Options.php
@@ -8,7 +8,10 @@ use ParaTest\Util\Str;
 use PHPUnit\TextUI\XmlConfiguration\Configuration;
 use PHPUnit\TextUI\XmlConfiguration\Loader;
 use RuntimeException;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputDefinition;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Process\Process;
 
 use function array_diff_key;
@@ -302,6 +305,190 @@ final class Options
         }
 
         return $options;
+    }
+
+    public static function setInputDefinition(InputDefinition $inputDefinition): void
+    {
+        $inputDefinition->setDefinition([
+            new InputOption(
+                'processes',
+                'p',
+                InputOption::VALUE_REQUIRED,
+                'The number of test processes to run.',
+                'auto'
+            ),
+            new InputOption(
+                'functional',
+                'f',
+                InputOption::VALUE_NONE,
+                'Run test methods instead of classes in separate processes.'
+            ),
+            new InputOption(
+                'no-test-tokens',
+                null,
+                InputOption::VALUE_NONE,
+                'Disable TEST_TOKEN environment variables. <comment>(default: variable is set)</comment>'
+            ),
+            new InputOption(
+                'help',
+                'h',
+                InputOption::VALUE_NONE,
+                'Display this help message.'
+            ),
+            new InputOption(
+                'coverage-clover',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Generate code coverage report in Clover XML format.'
+            ),
+            new InputOption(
+                'coverage-crap4j',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Generate code coverage report in Crap4J XML format.'
+            ),
+            new InputOption(
+                'coverage-html',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Generate code coverage report in HTML format.'
+            ),
+            new InputOption(
+                'coverage-php',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Serialize PHP_CodeCoverage object to file.'
+            ),
+            new InputOption(
+                'coverage-text',
+                null,
+                InputOption::VALUE_NONE,
+                'Generate code coverage report in text format.'
+            ),
+            new InputOption(
+                'coverage-xml',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Generate code coverage report in PHPUnit XML format.'
+            ),
+            new InputOption(
+                'coverage-test-limit',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Limit the number of tests to record for each line of code. Helps to reduce memory and size of ' .
+                'coverage reports.'
+            ),
+            new InputOption(
+                'max-batch-size',
+                'm',
+                InputOption::VALUE_REQUIRED,
+                'Max batch size (only for functional mode).',
+                0
+            ),
+            new InputOption(
+                'filter',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Filter (only for functional mode).'
+            ),
+            new InputOption(
+                'parallel-suite',
+                null,
+                InputOption::VALUE_NONE,
+                'Run the suites of the config in parallel.'
+            ),
+            new InputOption(
+                'passthru',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Pass the given arguments verbatim to the underlying test framework. Example: ' .
+                '--passthru="\'--prepend\' \'xdebug-filter.php\'"'
+            ),
+            new InputOption(
+                'passthru-php',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Pass the given arguments verbatim to the underlying php process. Example: --passthru-php="\'-d\' ' .
+                '\'zend_extension=xdebug.so\'"'
+            ),
+            new InputOption(
+                'whitelist',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Directory to add to the coverage whitelist.'
+            ),
+            new InputOption(
+                'phpunit',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'The PHPUnit binary to execute. <comment>(default: vendor/bin/phpunit)</comment>'
+            ),
+            new InputOption(
+                'runner',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Runner, WrapperRunner or SqliteRunner. <comment>(default: Runner)</comment>'
+            ),
+            new InputOption(
+                'bootstrap',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'The bootstrap file to be used by PHPUnit.'
+            ),
+            new InputOption(
+                'configuration',
+                'c',
+                InputOption::VALUE_REQUIRED,
+                'The PHPUnit configuration file to use.'
+            ),
+            new InputOption(
+                'group',
+                'g',
+                InputOption::VALUE_REQUIRED,
+                'Only runs tests from the specified group(s).'
+            ),
+            new InputOption(
+                'exclude-group',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Don\'t run tests from the specified group(s).'
+            ),
+            new InputOption(
+                'stop-on-failure',
+                null,
+                InputOption::VALUE_NONE,
+                'Don\'t start any more processes after a failure.'
+            ),
+            new InputOption(
+                'log-junit',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Log test execution in JUnit XML format to file.'
+            ),
+            new InputOption(
+                'colors',
+                null,
+                InputOption::VALUE_NONE,
+                'Displays a colored bar as a test result.'
+            ),
+            new InputOption(
+                'testsuite',
+                null,
+                InputOption::VALUE_OPTIONAL,
+                'Filter which testsuite to run'
+            ),
+            new InputArgument(
+                'path',
+                InputArgument::OPTIONAL,
+                'The path to a directory or file containing tests. <comment>(default: current directory)</comment>'
+            ),
+            new InputOption(
+                'path',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'An alias for the path argument.'
+            ),
+        ]);
     }
 
     /**
