+++ test/Functional/PHPUnitTest.php
@@ -372,153 +372,4 @@ final class PHPUnitTest extends FunctionalTestBase
         );
         $this->assertTestsPassed($proc, '2', '2');
     }
-
-    /**
-     * @return array<int, array<int, string|array<string, string>>>
-     */
-    /*
-    public function setsCoveragePhpDataProvider(): array
-    {
-        return [
-            [
-                ['coverage-html' => 'wayne'],
-                '',
-            ],
-            [
-                ['coverage-clover' => 'wayne'],
-                '',
-            ],
-            [
-                ['coverage-php' => 'notWayne'],
-                'notWayne',
-            ],
-            [
-                ['coverage-clover' => 'wayne', 'coverage-php' => 'notWayne'],
-                'notWayne',
-            ],
-        ];
-    }
-    */
-
-    /**
-     * @param array<string, string> $options
-     *
-     * @dataProvider setsCoveragePhpDataProvider
-     */
-    /*
-    public function testSetsCoveragePhp(array $options, string $coveragePhp): void
-    {
-        $phpUnit = new PHPUnit();
-        $c       = new ParaTestCommand($phpUnit);
-
-        $input = new ArrayInput([], $c->getDefinition());
-        foreach ($options as $key => $value) {
-            $input->setOption($key, $value);
-        }
-
-        $input->setArgument('path', '.');
-        $options = $phpUnit->getRunnerOptions($input);
-
-        if ($coveragePhp !== '') {
-            static::assertEquals($coveragePhp, $options['coverage-php']);
-        } else {
-            $dir = tempnam(sys_get_temp_dir(), 'paratest_');
-            static::assertStringStartsWith(dirname($dir), $options['coverage-php']);
-        }
-    }
-    */
-
-    /**
-     * @param array<string, array<string, int|string>> $options
-     * @param array<string, array<string, int|string>> $expected
-     *
-     * @dataProvider getRunnerOptionsDataProvider
-     */
-    /*
-    public function testGetRunnerOptions(array $options, array $expected): void
-    {
-        $phpUnit = new PHPUnit();
-        $c       = new ParaTestCommand($phpUnit);
-        $input   = new ArrayInput($options, $c->getDefinition());
-
-        $options = $phpUnit->getRunnerOptions($input);
-
-        // Note:
-        // 'coverage-php' contains a random, temporary string.
-        // has to be refactored to be testable but I'll leave that as a
-        // TODO
-        if (array_key_exists('coverage-php', $options)) {
-            unset($options['coverage-php']);
-        }
-
-        static::assertEquals($expected, $options);
-    }
-    */
-
-    /**
-     * @return array<string, array<string, array<string, array<int, string>|int|string>>>
-     */
-    /*
-    public function getRunnerOptionsDataProvider(): array
-    {
-        return [
-            'default' => [
-                'input' => [
-                    'path' => 'bar',
-                    '--processes' => '10',
-                ],
-                'expected' => [
-                    'path' => 'bar',
-                    'processes' => '10',
-                ],
-            ],
-            'accepts all defined options' => [
-                'input' => [
-                    'path' => 'bar',
-                    '--processes' => '10',
-                    '--functional' => 1,
-                    '--no-test-tokens' => 1,
-//                    '--help' => "",
-                    '--coverage-clover' => 'clover',
-                    '--coverage-crap4j' => 'xml',
-                    '--coverage-html' => 'html',
-                    '--coverage-text' => 'text',
-                    '--coverage-xml' => 'xml',
-                    '--max-batch-size' => '5',
-                    '--filter' => 'filter',
-                    '--parallel-suite' => 'parallel-suite',
-                ],
-                'expected' => [
-                    'path' => 'bar',
-                    'processes' => '10',
-                    'functional' => 1,
-                    'no-test-tokens' => 1,
-                    'coverage-clover' => 'clover',
-                    'coverage-crap4j' => 'xml',
-                    'coverage-html' => 'html',
-                    'coverage-text' => 'text',
-                    'coverage-xml' => 'xml',
-                    'max-batch-size' => '5',
-                    'filter' => 'filter',
-                    'parallel-suite' => 'parallel-suite',
-                ],
-            ],
-            "splits testsuite on ','" => [
-                'input' => [
-                    'path' => 'bar',
-                    '--processes' => '10',
-                    '--testsuite' => 't1,t2',
-                ],
-                'expected' => [
-                    'path' => 'bar',
-                    'processes' => '10',
-                    'testsuite' => [
-                        't1',
-                        't2',
-                    ],
-                ],
-            ],
-        ];
-    }
-    */
 }
